issue-72: Move index tracking into model
Keep the "code" logic in the model, and the "game" logic in the world.
