+++ src/MAKE_ATR
@@ -76,6 +76,7 @@ TLAVA  = t${LAVA}
 TLEVEL = t${LEVEL}
 TSPIKE = t${SPIKE}
 TWORLD = t${WORLD}
+TMODEL = t${MODEL}
 TCAMERA = t${CAMERA}
 
 # File Names, Tests - Raster
@@ -148,6 +149,9 @@ $(TSPIKE):
 $(TWORLD):
 	$(ATARI_CC) $(TEST_FLAGS) $(TESTS_MODEL_FOLDER)\$(TWORLD).c $(TEST_SRC) $(UNITY) -o $(TWORLD).ut
 
+$(TMODEL):
+	$(ATARI_CC) $(TEST_FLAGS) $(TESTS_MODEL_FOLDER)\$(TMODEL).c $(TEST_SRC) $(UNITY) -o $(TMODEL).ut
+
 $(TCAMERA):
 	$(ATARI_CC) $(TEST_FLAGS) $(TESTS_MODEL_FOLDER)\$(TCAMERA).c $(TEST_SRC) $(UNITY) -o $(TCAMERA).ut
 
@@ -204,13 +208,14 @@ $(TPLT_CLIP): $(PLT_CLIP).o
 
 
 # ----- Test Execution -------------------------------------------------------
-test: $(TBLOCK) $(TGEO) $(TLAVA) $(TLEVEL) $(TSPIKE) $(TWORLD) $(TCLR_SCN) $(TCLR_REG) $(TPLT_PXL) $(TPLT_HOR) $(TPLT_VRT) $(TPLT_LIN) $(TPLT_REC) $(TPLT_SQR) $(TPLT_TRI) $(TBMAP_8) $(TBMAP_16) $(TBMAP_32) $(TPLT_CHR) $(TPLT_STR) $(TBOUNDS) $(TPLT_CLIP)
+test: $(TBLOCK) $(TGEO) $(TLAVA) $(TLEVEL) $(TSPIKE) $(TWORLD) $(TMODEL) $(TCLR_SCN) $(TCLR_REG) $(TPLT_PXL) $(TPLT_HOR) $(TPLT_VRT) $(TPLT_LIN) $(TPLT_REC) $(TPLT_SQR) $(TPLT_TRI) $(TBMAP_8) $(TBMAP_16) $(TBMAP_32) $(TPLT_CHR) $(TPLT_STR) $(TBOUNDS) $(TPLT_CLIP)
 	$(TBLOCK).ut
 	$(TGEO).ut
 	$(TLAVA).ut
 	$(TLEVEL).ut
 	$(TSPIKE).ut
 	$(TWORLD).ut
+	$(TMODEL).ut
 	$(TCLR_SCN).ut
 	$(TCLR_REG).ut
 	$(TPLT_PXL).ut
